Add a check and buffer variable in SynchronizerImpl (#1733)
This commit adds a non-zero buffer check in the applyNextBlock method in the SynchronizerImpl class. This change helps prevent crashes by throwing an exception if the buffer is non-zero. Additionally, a buffer variable, buf_, has been added in the SynchronizerImpl class to aid in resolving the Logger crash issue (#1732). An additional short note (TODO) indicates this fix is temporary and should be removed once the root issue is fully addressed.